+++ include/neuron.h
@@ -6,16 +6,16 @@
 #include "connection.h"
 
 using namespace std;
-class Neuron{ 
+class Neuron{
 
-	double outputValue;
-	
-	double gradient;
+	long double outputValue;
+
+	long double gradient;
 
 	string transferFunction;
 
 	unsigned index;
- 	
+
  	// [0.0..1.0] overall net training rate
  	// 0.0 slow learner
  	// 0.2 medium learner
@@ -29,39 +29,37 @@ class Neuron{
 	vector<Connection> outputWeights;
 
 	//	---- transfer Functions hiperbolic tangent
-    static double transferFunctionTanH(double x);
+    static long double transferFunctionTanH(long double x);
 
-	static double transferFunctionTanHDerivative(double x);
+	static long double transferFunctionTanHDerivative(long double x);
 
 	//	---- transfer Functions sigmoid
-    static double transferFunctionSig(double x);
+    static long double transferFunctionSig(long double x);
 
-	static double transferFunctionSigDerivative(double x);
+	static long double transferFunctionSigDerivative(long double x);
 
 	//	---- transfer Functions step
-    static double transferFunctionStep(double x);
+    static long double transferFunctionStep(long double x);
 
-	static double transferFunctionStepDerivative();
-   
+	static long double transferFunctionStepDerivative();
+
     // sum of diff of weights of next layer
-    double sumDOW(const Layer &nextLayer) const;
+    long double sumDOW(const Layer &nextLayer) const;
 
 public:
 	Neuron(unsigned numberOutputs, unsigned index);
 
 	Neuron(unsigned numberOutputs, unsigned index, const string &transferFunction);
 
-	void setOutputValue(double value);
+	void setOutputValue(long double value);
 
-	double getOutputValue() const;
+	long double getOutputValue() const;
 
 	void feedForward(const Layer &prevLayer);
 
-	void calculateOutputGradients(double targetValue);
+	void calculateOutputGradients(long double targetValue);
 	void calculateHiddenGradients(const Layer &nextLayer);
-	void updateInputWeights(Layer &prevLayer, const double &eta, const double &alpha);
-	
-
+	void updateInputWeights(Layer &prevLayer, const long double &eta, const long double &alpha);
 };
 
 #endif
