+++ include/trainingdata.h
@@ -11,8 +11,8 @@ public:
     TrainingData(const string filename);
     bool isEof();
     void getTopology(vector<unsigned> &topology);
-    void getEta(double &eta);
-    void getMomentum(double &momentum);
+    void getEta(long double &eta);
+    void getMomentum(long double &momentum);
     void getTransferFunction(string &transferFunction);
 
     // Returns the number of input values read from the file:
